agents_agentic_email_assistant_langchain_openai/requirements.txt
streamlit
pandas
scikit-learn==1.3.2
joblib
langchain
openai
python-dotenv